--- a/performance/data/.gitignore
+++ b/performance/data/.gitignore
@@ -1,0 +1,2 @@
+*
+!Makefile