--- a/performance/data/.gitignore
+++ b/performance/data/.gitignore
@@ -1,2 +1,3 @@
 *
 !Makefile
+!requirements.txt