--- a/performance/data/requirements.txt
+++ b/performance/data/requirements.txt
@@ -1,0 +1,3 @@
+stdpopsim
+tskit
+bio2zarr