@jeromekelleher 2024-10-11
Fantastic, great job @Will-Tyler! I love the simplicity :+1: 

And excellent performance all round too, this is great!

It would be good to have a ``requirements.txt`` in the performance directory for the Python requirements. I think we can assume that bgzip, bcftools and pv are installed. I guess it would be good to document this in the comments of the script, just to be clear.

The simulated data is very easy though, and I guess we do want to provide some real data too. We should add some 1000G data also (but can do this as a follow-up).

It would also be nice to collect the top level stats in a data frame, and output results as CSV.